statistics: one more mkdirs fix
git-svn-id: https://arran.fi.muni.cz/svn/stanse/trunk@877 98efdb2a-0913-0410-9d53-f03ad6a1d54d

+++ src/cz/muni/stanse/statistics/DocumentToFileWriter.java
@@ -18,7 +18,11 @@ public final class DocumentToFileWriter {
 
     public static boolean write(final Document data, final String outFile) {
         final File pathName = new File(outFile);
-        pathName.getParentFile().mkdirs();
+	if (!pathName.getParentFile().exists() &&
+		!pathName.getParentFile().mkdirs()) {
+	    System.err.println("Can't create directory structure");
+	    return false;
+	}
         FileOutputStream os;
         try {
             os = new FileOutputStream(pathName);
